Defined Blockchain Dependant Hash
The Blockchain Dependant Hash (BloDHa) has been proposed earlier but not properly defined yet. This proposal increases the accuracy of the definition, adds a discussion of the topic and explains why it is a useful extension to an existing hash algorithm. 
The name Blodha can be confusing since Blodha is only an extension of existing hash algorithms, making the algorithm more modular.